Backend/requirements.txt
attrs==22.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.2.3
Flask-Cors==3.0.10
flask-expects-json==1.7.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pur==7.1.0
pyrsistent==0.19.3
pytz==2023.3
requests==2.30.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.15.0
